moved player classes into own sub folder
