+++ src/multiplayer/player/player_base.ts
@@ -1,6 +1,6 @@
 import type IMultiplayerMediator from "src/multiplayer/multiplayer_mediator";
 import type { MoveResult } from "src/omok_engine/move_status";
-import type IMove from "./i_move";
+import type IMove from "../i_move";
 
 export default abstract class PlayerBase{
     constructor(protected mediator: IMultiplayerMediator){
